requirements.txt
streamlit
langchain-groq
python-dotenv
requests
groq